Create processor for changing a crispr plan from "Attempt in Progress" to "Embryos Obtained"#484  - Includes refactoring to structure better the packages for state machine in plans.
