+++ impc_prod_tracker/core/src/main/java/org/gentar/biology/plan/engine/crispr/processors/AttemptInProgressProcessor.java
@@ -1,4 +1,4 @@
-package org.gentar.biology.plan.engine.processors.crispr;
+package org.gentar.biology.plan.engine.crispr.processors;
 
 import org.gentar.biology.plan.Plan;
 import org.gentar.biology.plan.engine.PlanStateSetter;
@@ -7,9 +7,9 @@ import org.gentar.statemachine.ProcessData;
 import org.springframework.stereotype.Component;
 
 @Component
-public class CrisprPlanAttemptInProgressProcessor extends AbstractProcessor
+public class AttemptInProgressProcessor extends AbstractProcessor
 {
-    public CrisprPlanAttemptInProgressProcessor(PlanStateSetter planStateSetter)
+    public AttemptInProgressProcessor(PlanStateSetter planStateSetter)
     {
         super(planStateSetter);
     }
