+++ impc_prod_tracker/core/src/main/java/org/gentar/biology/plan/starting_point/PlanStartingPoint.java
@@ -1,13 +1,11 @@
-package org.gentar.biology.plan.plan_starting_point;
+package org.gentar.biology.plan.starting_point;
 
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.gentar.BaseEntity;
-import org.gentar.biology.colony.Colony;
 import org.gentar.biology.outcome.Outcome;
 import org.gentar.biology.plan.Plan;
-import org.gentar.biology.specimen.Specimen;
 
 import javax.persistence.*;
 import java.io.Serializable;
