+++ impc_prod_tracker/core/src/test/java/org/gentar/statemachine/SystemEventsExecutorTest.java
@@ -3,8 +3,8 @@ package org.gentar.statemachine;
 import org.gentar.biology.plan.Plan;
 import org.gentar.biology.plan.attempt.AttemptTypes;
 import org.gentar.biology.plan.engine.PlanStateMachineResolver;
-import org.gentar.biology.plan.engine.events.CrisprProductionPlanEvent;
-import org.gentar.biology.plan.engine.state.CrisprProductionPlanState;
+import org.gentar.biology.plan.engine.crispr.CrisprProductionPlanEvent;
+import org.gentar.biology.plan.engine.crispr.CrisprProductionPlanState;
 import org.gentar.exceptions.SystemOperationFailedException;
 import org.gentar.exceptions.UserOperationFailedException;
 import org.gentar.test_util.PlanBuilder;
